LF: Refactor PreprocessorSpec test
Just moving stuff arround

CHANGELOG_BEGIN
CHANGELOG_END
