+++ ledger/participant-state/kvutils/src/test/suite/scala/com/daml/ledger/participant/state/kvutils/committer/transaction/validation/TransactionConsistencyValidatorSpec.scala
@@ -278,7 +278,7 @@ class TransactionConsistencyValidatorSpec extends AnyWordSpec with Matchers {
   private def create(
       contractId: Value.ContractId,
       signatories: Set[Ref.Party] = Set(aKeyMaintainer),
-      argument: TransactionBuilder.Value = aDummyValue,
+      argument: Value = aDummyValue,
       keyAndMaintainer: Option[(String, String)] = Some(aKey -> aKeyMaintainer),
   ): TransactionBuilder.Create =
     txBuilder.create(
